+++ misspy/hook.py
@@ -1,4 +1,13 @@
+# 互換性維持のためにしばらく残しておく
+
+import warnings
+
+
 class hook:
+    warnings.warn(
+        "misspy.hook.hook has been replaced by missspy.client.hook (missspy.hook) and will be obsolete after a certain period of time.",
+        DeprecationWarning,
+    )
     functions = {}
 
     def add(event, func):
@@ -24,7 +33,7 @@ class hook:
         Returns:
             bool: True
         """
-        
+
         hook.functions[event] = None
         return True
 
@@ -32,14 +41,14 @@ class hook:
         """reload
 
         Internally, it just executes remove and add.
-        
+
         Args:
             event (str): streaming event type. (view docs)
 
         Returns:
             bool: True
         """
-        
+
         func = hook.functions[event]
         hook.remove(event)
         hook.add(event, func)
